Added container width

+++ src/pages/app/MyCompany/MyCompanyNew.vue
@@ -1,5 +1,6 @@
 <template>
   <div class="md-layout">
+  <vue-element-loading :active="isLoading" spinner="ring" color="#FF547C"/>
     <div class="md-layout-item md-small-size-30 md-medium-size-30 md-large-size-20">
       <md-card class="md-card-profile">
         <md-card-content>
@@ -314,7 +315,7 @@
                 <LineChart
                   v-if='!isMonthly&&getChartNumberOfEventsPerYear'
                   key="username-input"
-                  classStyle="max-height: 130px;  border-radius: 5px; box-shadow: 0px 2px 9px 0 rgba(0, 0, 0, 0.31); background-image: linear-gradient(322deg, #4d9b51, #62b766);"
+                  classStyle="max-height: 150px;  border-radius: 5px; box-shadow: 0px 2px 9px 0 rgba(0, 0, 0, 0.31); background-image: linear-gradient(322deg, #4d9b51, #62b766);"
                   id="number_of_events_chart"
                   width="350"
                   height="150"
@@ -324,7 +325,7 @@
                 <LineChart
                   v-else
                   key="email-input"
-                  classStyle="max-height: 130px;  border-radius: 5px;box-shadow: 0px 2px 9px 0 rgba(0, 0, 0, 0.31);background-image: linear-gradient(322deg, #4d9b51, #62b766);"
+                  classStyle="max-height: 150px;  border-radius: 5px;box-shadow: 0px 2px 9px 0 rgba(0, 0, 0, 0.31);background-image: linear-gradient(322deg, #4d9b51, #62b766);"
                   id="number_of_events_chart_monthly"
                   width="350"
                   height="150"
@@ -374,14 +375,14 @@
             <div class='chart-title'>
             <div class="title text-bold">Number of events</div>
             <div class="company-button-filter-block">
-              <ButtonDiv text='Yearly' class='button-filter'  :onClick='onChangeFilterToEarly'/>
-              <ButtonDiv text='Monthly' class='button-filter'   :onClick='onChangeFilter'/>
+              <ButtonDiv text='Yearly' class='button-filter' secondClass='chenge-filter' :classChenger='!showFilter'  :onClick='onChangeFilterToEarly'/>
+              <ButtonDiv text='Monthly' class='button-filter' secondClass='chenge-filter' :classChenger='showFilter'    :onClick='onChangeFilter'/>
             </div>
           </div>
           <div><span class='info-chat-value'>{{numberOfEvents}}</span><span class='info-chart'>{{`This year of ${currentYear}`}}</span></div>
         </md-card-content>
       </div>
-      <md-card>
+      <md-card class='margin-block-for-card'>
       <div class="logo-block">
           <div class="event-planer-logo partisipation-logo">
             <md-icon class="company-logo ">how_to_reg</md-icon>
@@ -406,11 +407,11 @@
           </div>
         </md-card-content>
       </md-card>
-      <md-card>
+      <md-card class='margin-block-for-card'>
       <div class="logo-block">
          <LineChart
             v-if='getChartEventPerEmployee'
-            classStyle="max-height: 130px; border-radius: 5px;  box-shadow: 0px 2px 9px 0 rgba(0, 0, 0, 0.31);  background-image: linear-gradient(322deg, #c3255b, #ea3c77);"
+            classStyle="max-height: 150px; border-radius: 5px;  box-shadow: 0px 2px 9px 0 rgba(0, 0, 0, 0.31);  background-image: linear-gradient(322deg, #c3255b, #ea3c77);"
             id="number_of_participants_chart"
             width="350"
             height="150"
@@ -423,7 +424,7 @@
         </md-card-content>
       </md-card>
     </div>
-    <div class="md-layout-item md-size-33">
+    <div class="md-layout-item md-size-33 ">
       <md-card>
        <div class="logo-block">
           <div class="event-planer-logo rate-logo">
@@ -453,11 +454,11 @@
           </div>
         </md-card-content>
       </md-card>
-      <md-card>
+      <md-card class='margin-block-for-card'>
       <div class="logo-block">
           <LineChart
             v-if='getChartEventsPerCategory'
-            classStyle="max-height: 130px; border-radius: 5px;  box-shadow: 0px 2px 9px 0 rgba(0, 0, 0, 0.31);  background-image: linear-gradient(322deg, #4d9b51, #62b766);"
+            classStyle="max-height: 150px; border-radius: 5px;  box-shadow: 0px 2px 9px 0 rgba(0, 0, 0, 0.31);  background-image: linear-gradient(322deg, #4d9b51, #62b766);"
             id="event_vs_category"
             width="350"
             height="150"
@@ -478,9 +479,11 @@
   import {mapGetters, mapMutations, mapState} from 'vuex';
   import Chart from 'chart.js';
   import auth from '@/auth';
+   import swal from "sweetalert2";
 
   //COMPONENTS
   import { Tabs, NavTabsCard } from "@/components";
+  import VueElementLoading from 'vue-element-loading';
   import Select from '@/components/Select/Select.vue'
   import MyCompanyDashboard from "src/pages/app/MyCompany/MyCompanyDashboard.vue";
   import MyCompanyApprovals from "src/pages/app/MyCompany/MyCompanyApprovals.vue";
@@ -523,15 +526,17 @@ const currentYear=new Date().getFullYear()
       LineIndicator,
       Datepicker,
       "select-common":Select,
-      LineChart
-    },    
-    mounted(){      
+      LineChart,
+      VueElementLoading
+    }   
+    ,    
+     mounted(){      
+
        const options = {          
           types: ['geocode']
         }
         let input = document.getElementById('branch_address_search')
-        let autocomplete = new google.maps.places.Autocomplete(input, options)
-        console.log(autocomplete.description)
+        let autocomplete = new google.maps.places.Autocomplete(input, options)        
         this.auth.currentUser(this, true, function() {
           this.$store.dispatch("user/getUserFromApi");
           this.$store.dispatch("user/getIndustry");   
@@ -583,6 +588,7 @@ const currentYear=new Date().getFullYear()
                   
               }],
               xAxes: [{
+                barThickness:20,
                 ticks: {
                   beginAtZero:true,
                   fontColor:"white",
@@ -618,7 +624,8 @@ const currentYear=new Date().getFullYear()
         formSwitcher:'',
         duration:[],
         isEnabled: true,
-        currentYear:currentYear
+        currentYear:currentYear,
+        isLoadingScreen:true
       }
     },
     computed:{
@@ -633,6 +640,9 @@ const currentYear=new Date().getFullYear()
         participants:'user/getChartParticipantsPerEvent',
         rate:'user/getChartSatisfactionRate'
       }),
+     isLoading(){        
+       return !Boolean(this.charts)
+      },
       numberOfEvents(){
         if(this.charts.numberOfEventsPerYear){
           const charts=this.charts.numberOfEventsPerYear
@@ -768,6 +778,7 @@ const currentYear=new Date().getFullYear()
       }
     }
     ,
+
     methods: {
       onUpdateFocus(newValue) {
         this.inputFocus = newValue;
@@ -861,14 +872,26 @@ const currentYear=new Date().getFullYear()
         console.log('@')
         document.getElementById('company-avatar-upload').click()
       },
-      deleteAvatar(id){        
-       new CustomerFile({id: id}).delete().then(result => {
+      deleteAvatar(id){ 
+        
+        swal({
+          title: "Are you sure want to delete company logo?",          
+          showCancelButton: true,
+          confirmButtonClass: "md-button md-success",
+          cancelButtonClass: "md-button md-danger",
+          confirmButtonText: "Yes, delete it!",
+          buttonsStyling: false
+        }).then(result => {
+           new CustomerFile({id: id}).delete().then(result => {
           this.$store.dispatch("user/getUserFromApi"); 
          
         })
           .catch((error) => {
             console.log(error);
           });
+            }).catch(() => {
+               console.log(error);
+            });
       },
     
     }, 
@@ -908,6 +931,7 @@ const currentYear=new Date().getFullYear()
       }
     } 
     .value {
+      width: 30%;
       color:#292929!important;
     }
     p {
@@ -979,8 +1003,16 @@ const currentYear=new Date().getFullYear()
     color: #c6c6c6;
     margin-left: 5px;
   }
-  .text-bold {
-    font-weight: 600;
+  .text-bold {  
+      font-family: 'Roboto';
+      font-size: 12px;
+      font-weight: 300;
+      font-style: normal;
+      font-stretch: normal;
+      line-height: 2;
+      letter-spacing: normal;
+      text-align: left;
+      color: #292929;
   }
   .info-text-size {
     font-size: 14px;
@@ -1015,6 +1047,16 @@ const currentYear=new Date().getFullYear()
     padding: 0px 6px;
     background: #62b766;
     color: white;
+    border-radius: 5px;
+    cursor: pointer;
+    margin:5px;
+    width: 90px;
+  }
+  .chenge-filter{
+    background:#f0f0f0;
+    text-align: center;
+    padding: 0px 6px;
+    color: black;
     border-radius: 5px;
     cursor: pointer;
     margin:5px;
@@ -1072,6 +1114,7 @@ const currentYear=new Date().getFullYear()
     position: absolute;
     top: -40px;
     left: 20px;
+    z-index: 2; 
   }
   .indicator-event-type-title-rate{
     font-size: 0.85rem;
@@ -1126,7 +1169,7 @@ const currentYear=new Date().getFullYear()
   justify-content: space-between;
   align-items: flex-end;
   padding: 0px 15px;
-  margin: -20px 0px 20px 0px;
+  margin: -20px 0px 0px 0px;
   .event-planer-logo {
     background: #eb3e79;
     width: 64px;
@@ -1151,4 +1194,13 @@ const currentYear=new Date().getFullYear()
     color: #000000;
   }
 }
+.margin-block-for-card{
+  margin: 40px 0 !important;
+}
 </style>
+
+
+
+
+        
+
